# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brasileirão API",
        "description": "API REST para o sistema do Campeonato Brasileiro Série A",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/admin/teams": {
            "get": {
                "tags": [
                    "Admin - Teams"
                ],
                "summary": "Listar todos os times",
                "operationId": "764f07b4be9872117aef38607cdfb26b",
                "responses": {
                    "200": {
                        "description": "Lista de times"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Teams"
                ],
                "summary": "Criar time",
                "operationId": "225d7a32d9d6b773a1a5a9e183e4c42d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "short_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Flamengo"
                                    },
                                    "short_name": {
                                        "type": "string",
                                        "example": "FLA",
                                        "maxLength": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Time criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/teams/{id}": {
            "get": {
                "tags": [
                    "Admin - Teams"
                ],
                "summary": "Detalhe do time",
                "operationId": "2382f6da05826836ee760173ed984cc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Time encontrado"
                    },
                    "404": {
                        "description": "Time não encontrado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Teams"
                ],
                "summary": "Editar time",
                "operationId": "683dfa490cef21f8f47d42dbc1735f12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "short_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Flamengo"
                                    },
                                    "short_name": {
                                        "type": "string",
                                        "example": "FLA",
                                        "maxLength": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Time atualizado"
                    },
                    "404": {
                        "description": "Time não encontrado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Teams"
                ],
                "summary": "Excluir time",
                "operationId": "566af7ebe9ef0e66ed293914a8d523ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Time removido"
                    },
                    "404": {
                        "description": "Time não encontrado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}